Switch number type from f64 to f32 throughout core
Replaces all uses of f64 with f32 for the Number type in op.rs, type.rs, and value.rs. Updates parsing, formatting, and code generation to use f32, ensuring consistency in floating-point representation.

+++ core/src/type.rs
@@ -58,7 +58,7 @@ impl Node for Type {
     fn compile(&self, ctx: &mut Compiler) -> Option<String> {
         Some(
             match self.type_infer(ctx)? {
-                Type::Number => "f64",
+                Type::Number => "f32",
                 Type::Integer
                 | Type::Bool
                 | Type::String
